+++ src/main/java/com/threestar/trainus/domain/metadata/dto/ProfileMetadataResponseDto.java
@@ -3,6 +3,6 @@ package com.threestar.trainus.domain.metadata.dto;
 public record ProfileMetadataResponseDto(
 	Long userId,
 	Integer reviewCount,
-	Float rating
+	Double rating
 ) {
 }
